Merged PR 2515: Updating EXM Email Template Branding
Updating branding in Shared Repo - will move to Platform repo in a separate task (5207)

For testing you can use the Platform !2516 to test. But there are no Platform changes for this task.

The purpose of this PR is to remove the Habitat Home branding from the email template we add for EXM. There is no Lighthouse logo added. It is left "Unbranded - No logo, just colors" for all EXM emails so they don't need to be updated by SE's.

You can access the "Branded Email" template in EXM by creating either a "Regular email campaign" or a "Automated email campaign" and it is listed last on the second level in the list of EXM templates.

Related work items: #5038